Prepared the project for 3.3.10 release. Fixed SSE .Sum() for double[]

+++ HPCsharp/SumParallel.cs
@@ -3,7 +3,7 @@
 // TODO: Implement aligned SIMD sum, since memory alignment is critical for SIMD instructions. So, do scalar first until we are SIMD aligned and then do SIMD, followed by more scarlar to finish all
 //       left over elements that are not SIMD size divisible. First simple step is to check alignment of SIMD portion of the sum.
 // TODO: Contribute to Sum C# stackoverflow page, since nobody considered overflow condition and using a larger range values for sum. Also, to https://stackoverflow.com/questions/9987560/comparing-sum-methods-in-c-sharp which
-//       asks for a faster implementation, so we can point to SSE and multi-core implemenations.
+//       asks for a faster implementation, so we can point to SSE and multi-core implemenations. Also, contribute to https://stackoverflow.com/questions/26548756/looking-for-a-faster-way-to-sum-arrays-in-c-sharp
 // TODO: Develop a method to split an array on a cache line (64 byte) boundary. Make it public, since it will be useful in many cases.
 // TODO: Implement a for loop instead of divide-and-conquer, since they really accomplish the same thing, and the for loop will be more efficient and easier to make cache line boundary divisible.
 //       Combining will be slightly harder, but we could create an array of sums, where each task has its own array element to fill in, then we combine all of the sums at the end serially.
@@ -426,7 +426,7 @@ namespace HPCsharp
             var sumVector = new Vector<double>();
             int sseIndexEnd = l + ((r - l + 1) / Vector<double>.Count) * Vector<double>.Count;
             int i;
-            for (i = l; i < sseIndexEnd; i += Vector<float>.Count)
+            for (i = l; i < sseIndexEnd; i += Vector<double>.Count)
             {
                 var inVector = new Vector<double>(arrayToSum, i);
                 sumVector += inVector;
@@ -679,7 +679,7 @@ namespace HPCsharp
             double sumRight = 0;
 
             Parallel.Invoke(
-                () => { sumLeft = SumSseParInner(arrayToSum, l, m); },
+                () => { sumLeft  = SumSseParInner(arrayToSum, l,     m); },
                 () => { sumRight = SumSseParInner(arrayToSum, m + 1, r); }
             );
             // Combine left and right results
